Split decoder tests into two projects
